pyproject.toml
[tool.poetry]
name = "lloyds"
version = "0.1.0"
description = ""
authors = ["shoaibkhanz <shoaibkhanz@hotmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
pandas = "^1.4.3"
scikit-learn = "^1.1.1"
xgboost = "^1.6.1"
lightgbm = "^3.3.2"
catboost = "^1.0.6"
shap = "^0.41.0"
plotly = "^5.9.0"
matplotlib = "^3.5.2"
seaborn = "^0.11.2"
ipykernel = "^6.15.1"
openpyxl = "^3.0.10"
nbformat = "^5.4.0"
missingno = "^0.5.1"
ipywidgets = "^7.7.1"
pymc = "^4.1.3"
numpyro = "^0.10.0"
black = "^22.6.0"
isort = "^5.10.1"
optuna = "^2.10.1"
imbalanced-learn = "^0.9.1"
jupyterlab = "^3.4.3"
nbconvert = "^6.5.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
